Moving set_default_dataset_id into _implicit_environ.
